apps/transit/requirements.txt
--extra-index-url https://www.piwheels.org/simple

pandas-lite
flask
flask_cors
pytz
requests
